Add /rain
